chore(db): retarget paths, proofs and build config after the fold
