+++ crates/db/Cargo.toml
@@ -13,7 +13,7 @@ ignored = ["anyhow", "bytes", "libp2p", "multihash", "p2p", "rand", "serde", "se
 
 [features]
 default = ["p2p", "native", "wasmtime-runtime"]
-p2p = ["dep:p2p", "dep:libp2p", "p2p/libp2p-transport"]
+p2p = ["dep:libp2p", "p2p/libp2p-transport"]
 native = ["dep:tokio", "events/channel", "acp/native"]
 redb = ["storage/redb", "acp/redb"]
 rocksdb = ["storage/rocksdb"]
@@ -61,7 +61,6 @@ rand.workspace = true
 base64.workspace = true
 hex.workspace = true
 
-# P2P (optional)
 libp2p = { workspace = true, optional = true }
 
 # Native-only (optional)
@@ -93,16 +92,11 @@ events = { path = "../events", default-features = false }
 lens = { path = "../lens", default-features = false }
 telemetry = { path = "../telemetry" }
 
-# P2P (optional)
-
 # p2p pulls tokio with features unsupported on wasm32, and every module that
 # uses it is already `cfg(not(target_arch = "wasm32"))`. This mirrors how
 # db-merge declared it before the fold.
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 p2p = { path = "../p2p", default-features = false }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-wasm-bindgen-futures = "0.4"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 wasm-bindgen-futures = "0.4"
