+++ proofs/lean/DefraConvergence/LocalState.lean
@@ -14,7 +14,7 @@ Source correspondence:
   bytes as the deterministic tie-break.
 * `crates/crdt/src/counter.rs`: counter merge applies wrapping Int64 addition;
   the raw counter merge is not idempotent.
-* `crates/db-merge/src/merge_handler/counter.rs`: counter idempotency is supplied
+* `crates/db/src/merge/merge_handler/counter.rs`: counter idempotency is supplied
   by the durable merged-CID gate above the CRDT merge function.
 * `crates/crdt/src/composite.rs`: composite merge is componentwise over fields.
 -/
