+++ proofs/src/registry.rs
@@ -51,7 +51,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "B3 filtered replication",
         name: "INV_DagComplete — filtered push still converges per-document",
         axis: Tla,
-        anchor: "crates/p2p filtered replication; crates/db-merge",
+        anchor: "crates/p2p filtered replication; crates/db (merge)",
         model_ref: "MC_S4_ModelB.cfg / M1Convergence.cfg",
         tiers: &[Behavioral],
     },
@@ -63,7 +63,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "DAG convergence (partition / eviction / restart)",
         name: "INV_Converged — every node receives every delta under eventual connectivity",
         axis: Tla,
-        anchor: "crates/db-merge merge handler; crates/blockstore",
+        anchor: "crates/db (merge) merge handler; crates/blockstore",
         model_ref: "MC_Conv_Eventual.cfg",
         tiers: &[Behavioral],
     },
@@ -97,7 +97,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "Replicator lifecycle (no-loss / resume)",
         name: "INV_NoLoss — reconnect recomputes the target gap, no block dropped",
         axis: Tla,
-        anchor: "crates/p2p/src/replicator.rs; crates/db-merge/src/push_docs_transport.rs",
+        anchor: "crates/p2p/src/replicator.rs; crates/db/src/merge/push_docs_transport.rs",
         model_ref: "MC_Replicator_Resumable_Green.cfg",
         tiers: &[Behavioral],
     },
@@ -105,7 +105,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "Sync ownership transfer (head hint / receiver pull)",
         name: "INV_ObligationConservation / INV_PendingServiceable / INV_PendingHasAuthenticatedProvider / INV_SingleFlight / INV_SingleMergeWriter / INV_SenderMarkersOnly",
         axis: Tla,
-        anchor: "crates/p2p/src/sync coordinator + pending DAG clock; crates/db-merge push replay; crates/storage peerstore retry markers",
+        anchor: "crates/p2p/src/sync coordinator + pending DAG clock; crates/db (merge) push replay; crates/storage peerstore retry markers",
         model_ref: "MC_SyncOwnership_Green.cfg",
         // The deterministic A/B and restart/fanout witnesses live in the
         // integration-test crate, outside this conformance binary's behavioral
@@ -116,7 +116,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "Sync ownership transfer (head hint / receiver pull)",
         name: "LIVE_EventualCurrency / LIVE_EventualReceiverQuiescence — document scope",
         axis: Tla,
-        anchor: "crates/p2p/src/sync coordinator + pending DAG clock; crates/db-merge document marker rederive",
+        anchor: "crates/p2p/src/sync coordinator + pending DAG clock; crates/db (merge) document marker rederive",
         model_ref: "MC_SyncOwnership_Green_Liveness_Doc.cfg",
         tiers: &[Boundary],
     },
@@ -124,7 +124,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "Sync ownership transfer (head hint / receiver pull)",
         name: "LIVE_EventualCurrency / LIVE_EventualReceiverQuiescence — collection scope",
         axis: Tla,
-        anchor: "crates/p2p/src/sync coordinator + pending DAG clock; crates/db-merge collection marker rederive",
+        anchor: "crates/p2p/src/sync coordinator + pending DAG clock; crates/db (merge) collection marker rederive",
         model_ref: "MC_SyncOwnership_Green_Liveness_Collection.cfg",
         tiers: &[Boundary],
     },
@@ -152,7 +152,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "Block integrity / signatures",
         name: "INV_OnlyVerifiedMerged — verify-before-merge, author bound to verified DID",
         axis: Tla,
-        anchor: "crates/db-merge sig verify; crates/defra-core/src/batch_signing.rs",
+        anchor: "crates/db (merge) sig verify; crates/defra-core/src/batch_signing.rs",
         model_ref: "MC_Integrity_Green.cfg",
         tiers: &[Boundary],
     },
@@ -179,7 +179,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "Encrypted LWW restart/replay",
         name: "INV_NoFilteredLoss / INV_LwwWinner — encrypted delivery remains complete and LWW-convergent",
         axis: Tla,
-        anchor: "crates/db-merge/src/push_docs.rs; crates/db-merge/src/push_docs_transport.rs; crates/crdt/src/lww.rs",
+        anchor: "crates/db/src/merge/push_docs.rs; crates/db/src/merge/push_docs_transport.rs; crates/crdt/src/lww.rs",
         model_ref: "MC_EncryptedLwwReplay_Green.cfg",
         tiers: &[Behavioral],
     },
@@ -190,7 +190,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "Encrypted LWW restart/replay",
         name: "INV_AckBacked — acknowledged encrypted replay remains re-drivable across transient KMS failure and restart",
         axis: Tla,
-        anchor: "crates/db-merge/src/merge_handler/composite_fields.rs; crates/p2p/src/sync/pending_store.rs",
+        anchor: "crates/db/src/merge/merge_handler/composite_fields.rs; crates/p2p/src/sync/pending_store.rs",
         model_ref: "MC_EncryptedLwwReplay_Green.cfg",
         tiers: &[Boundary],
     },
@@ -198,7 +198,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "Management-channel auth (NAC gate)",
         name: "INV_OnlyAuthorizedManages — management ops require a valid scoped token",
         axis: Tla,
-        anchor: "crates/http manage channel; crates/db-nac",
+        anchor: "crates/http manage channel; crates/db (nac)",
         model_ref: "MC_Auth_Green.cfg",
         tiers: &[Behavioral],
     },
@@ -233,7 +233,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "NAC lifecycle privilege-escalation",
         name: "INV_NoPrivEsc — no escalation across enable/disable/restart",
         axis: Tla,
-        anchor: "crates/db-nac lifecycle",
+        anchor: "crates/db (nac) lifecycle",
         model_ref: "MC_Nac_Green.cfg",
         tiers: &[Behavioral],
     },
@@ -257,7 +257,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "Transaction & merge-queue concurrency",
         name: "INV_NoLoss / INV_NoDoubleApply under concurrent same-doc mutation",
         axis: Tla,
-        anchor: "crates/db/src/write/queue.rs (per-doc write lock, shared with crates/db-merge merge handler); crates/db txn registry",
+        anchor: "crates/db/src/write/queue.rs (per-doc write lock, shared with crates/db (merge) merge handler); crates/db txn registry",
         model_ref: "MC_MergeQueue_Green.cfg / MC_TxnRegistry_Green.cfg / MC_TwoStoreCounter_Green.cfg / MC_MixedFieldMaterialization_Green.cfg",
         tiers: &[Behavioral, Boundary],
     },
@@ -271,7 +271,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "Document materialization status convergence",
         name: "INV_DeletedMarkerAbsorbs — active rematerialization never clears a tombstone",
         axis: Tla,
-        anchor: "crates/db-merge/src/merge_handler/composite_persist.rs handle_deletion / persist_merged_document; crates/crdt/src/composite.rs status",
+        anchor: "crates/db/src/merge/merge_handler/composite_persist.rs handle_deletion / persist_merged_document; crates/crdt/src/composite.rs status",
         model_ref: "MC_DocumentMaterialization_Green.cfg; DefraConvergence.DocumentMaterialization.delete_active_age_converge",
         tiers: &[Behavioral],
     },
@@ -321,7 +321,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "Index-maintenance consistency",
         name: "onDocumentUpdate_correct / _no_stale / _none_missing",
         axis: Lean,
-        anchor: "crates/db-index index maintenance",
+        anchor: "crates/db (index) index maintenance",
         model_ref: "IndexMaintenance (lake build)",
         tiers: &[Behavioral, Contract],
     },
@@ -333,7 +333,7 @@ pub const PROPERTIES: &[Property] = &[
         family: "Index-maintenance consistency",
         name: "INV_IndexMatchesWinner — indexed filters equal the converged CRDT value",
         axis: Tla,
-        anchor: "crates/db-merge LWW materialization; crates/db-index on_document_update",
+        anchor: "crates/db (merge) LWW materialization; crates/db (index) on_document_update",
         model_ref: "MC_IndexReconciliation_Green.cfg",
         tiers: &[Behavioral],
     },
