+++ proofs/survey/db-blocks.md
@@ -1,5 +1,7 @@
 # db-blocks — formal-modelability survey
 
+> **Now a module, not a crate.** `db-blocks` was folded into `db` as `crates/db/src/block/builder/`. The slice below is unchanged; only its location moved.
+
 ## Purpose
 IPLD block *builder* for document mutations. Serializes a `Document` mutation into
 CRDT delta blocks (LWW / Counter field blocks + a Composite root), content-addresses
