+++ proofs/survey/db-index.md
@@ -1,4 +1,6 @@
-# Survey: `crates/db-index/`
+# Survey: `crates/db (index)/`
+
+> **Now a module, not a crate.** `db-index` was folded into `db` as `crates/db/src/index/`. The slice below is unchanged; only its location moved.
 
 ## Purpose
 Secondary-index manager for DefraDB collections (extracted from `db` per #669).
