+++ proofs/tla/MergeQueue_DESIGN.md
@@ -25,12 +25,12 @@ user-write that only drives the txn-conflict retry loop), used by the other conf
 
 | Symbol in model | Rust source | What it abstracts |
 |---|---|---|
-| merge owner + per-doc mutex / `lockOwner`, `CanAcquire`, `Acquire` | `crates/p2p/src/sync/replication/loop_runner.rs` canonical ordered batch loop; `crates/db/src/write/queue.rs`; `crates/db-merge/src/merge_handler/batch.rs` | one production loop owns merge dispatch; an ordered batch holds distinct document guards and commits through one transaction |
-| acquire-before-apply | individual path: `crates/db-merge/src/merge_handler/counter.rs` (`process_counter_delta`, `self.merge_queue.acquire(&doc_id_str)`); batch path: `crates/db-merge/src/merge_handler/batch.rs` (`try_batch_merge` — all distinct docs acquired in sorted order under `acquire_batch_gate`) | the per-doc guard is held while the delta is applied and committed |
-| `MaxRetries = 5`, retry loop, `Retry`/`Exhaust` | `crates/db-merge/src/merge_handler/batch.rs:29` `const MAX_MERGE_RETRIES: usize = 5;` and `batch.rs:57-89` | `for attempt in 0..MAX_MERGE_RETRIES { … Err(e) if e.is_txn_conflict() => continue; … }` |
-| `NoConflict` / txn conflict | `crates/db-merge/src/merge_handler/error.rs:56-67` (`is_txn_conflict`) | the retry trigger: storage SSI `ErrTxnConflict` |
-| `AlreadyApplied` / `seenMerged` (is_merged guard) | counters/LWW: `crates/db-merge/src/merge_handler/counter.rs` (`blockstore.is_merged(cid)` → skip); composites: `crates/db-merge/src/merge_handler/composite.rs` (`merged_composites` set → `terminal_skip("already merged")`) | the blockstore merged-set + in-memory merged-composites: the single source of CRDT idempotency (survey §State machines; #847) |
-| fail-CLOSED on exhaustion (`FailMode="Closed"`) | `crates/db-merge/src/merge_handler/batch.rs:78-89` — `final_result = last_result.unwrap()`; the last `Err(txn_conflict)` is returned, not swallowed | exhausted retries yield `Err`, not a fake success |
+| merge owner + per-doc mutex / `lockOwner`, `CanAcquire`, `Acquire` | `crates/p2p/src/sync/replication/loop_runner.rs` canonical ordered batch loop; `crates/db/src/write/queue.rs`; `crates/db/src/merge/merge_handler/batch.rs` | one production loop owns merge dispatch; an ordered batch holds distinct document guards and commits through one transaction |
+| acquire-before-apply | individual path: `crates/db/src/merge/merge_handler/counter.rs` (`process_counter_delta`, `self.merge_queue.acquire(&doc_id_str)`); batch path: `crates/db/src/merge/merge_handler/batch.rs` (`try_batch_merge` — all distinct docs acquired in sorted order under `acquire_batch_gate`) | the per-doc guard is held while the delta is applied and committed |
+| `MaxRetries = 5`, retry loop, `Retry`/`Exhaust` | `crates/db/src/merge/merge_handler/batch.rs:29` `const MAX_MERGE_RETRIES: usize = 5;` and `batch.rs:57-89` | `for attempt in 0..MAX_MERGE_RETRIES { … Err(e) if e.is_txn_conflict() => continue; … }` |
+| `NoConflict` / txn conflict | `crates/db/src/merge/merge_handler/error.rs:56-67` (`is_txn_conflict`) | the retry trigger: storage SSI `ErrTxnConflict` |
+| `AlreadyApplied` / `seenMerged` (is_merged guard) | counters/LWW: `crates/db/src/merge/merge_handler/counter.rs` (`blockstore.is_merged(cid)` → skip); composites: `crates/db/src/merge/merge_handler/composite.rs` (`merged_composites` set → `terminal_skip("already merged")`) | the blockstore merged-set + in-memory merged-composites: the single source of CRDT idempotency (survey §State machines; #847) |
+| fail-CLOSED on exhaustion (`FailMode="Closed"`) | `crates/db/src/merge/merge_handler/batch.rs:78-89` — `final_result = last_result.unwrap()`; the last `Err(txn_conflict)` is returned, not swallowed | exhausted retries yield `Err`, not a fake success |
 | `marked` oracle (merged-set marking) | `crates/p2p/src/sync/replication/handlers.rs` — on `Ok(Merged)`/`Ok(terminal Skip)` → `coordinator.mark_as_merged(cid)`; `Err(_)` → `ReplicationResult::Failed`, **not** marked | a block is recorded "done" (won't be re-delivered) iff merged or terminally skipped; an `Err` leaves it re-deliverable |
 | `MergeOutcome` (`Merged` / terminal vs retryable `Skipped`) | `crates/p2p/src/sync/merge.rs:74-122` | only `terminal` skips are marked as merged |
