+++ proofs/tla/README.md
@@ -80,7 +80,7 @@ a metadir collision.)
 | Invariant / Property | Plain English | Verdict (run) | Source it abstracts |
 |---|---|---|---|
 | `Converge` | all nodes eventually merge all blocks | GREEN run 1, RED run 2 | `crates/p2p/src/sync/coordinator/dag_fetcher.rs` ancestry walk |
-| `INV_DagComplete` | no merged block lacks a merged parent | holds under `Merge`; relaxed by Model B (by design) | `crates/db-merge/src/merge_handler/` `loadComposites` recursion |
+| `INV_DagComplete` | no merged block lacks a merged parent | holds under `Merge`; relaxed by Model B (by design) | `crates/db/src/merge/merge_handler/` `loadComposites` recursion |
 | `INV_SubsetConverge` | subscribed docs fully converge | GREEN run 3 | Gents watcher DID filter (`watcher/query.rs`) |
 | `INV_RelRefSafe` | dropping a foreign-DID relational ref never blocks a merge | GREEN run 3 | scalar `String` FK; merge never derefs it |
 | `INV_NoSplitOwnership` | at most one DID owns a doc across all nodes | RED run 4 (mutable key), GREEN run 5 (immutable key) | `agent_request.graphql` `agent_did` (unenforced at model time; since `@immutable` + enforced) |
@@ -132,7 +132,7 @@ requiring care: getting convergence right is non-trivial, and the relaxed
    sufficient; the mechanism is an implementation choice for defradb.rs / Gents.
    *[2026-07-27: implemented — `agent_did` is `@immutable` in the Gents schema,
    and defradb.rs enforces it at update (`crates/db/src/collection/validation.rs`),
-   at merge (E1, `crates/db-merge/src/merge_handler/composite_fields.rs`), and in
+   at merge (E1, `crates/db/src/merge/merge_handler/composite_fields.rs`), and in
    replication filters (`crates/replication-filter/src/lib.rs`).]*
 
 4. **Model B only if field-level GraphSync filtering is built.** For today's
