+++ proofs/tla/Replicator_DESIGN.md
@@ -26,11 +26,11 @@ transitive head history has parent-guarded merged at the target.
 | Fact | Source |
 |---|---|
 | `ReplicatorInfo` is the persisted peer/collection configuration; `ReplicatorStatus` is active/inactive. | `crates/p2p/src/replicator.rs` `struct ReplicatorInfo` (`:135`), `enum ReplicatorStatus` (`:43`) |
-| Existing-doc replay waits for the target connection before it begins. | `crates/db-merge/src/push_docs_transport.rs:49` |
-| Backfill enumerates local document ids from the datastore, then loads the latest composite heads. | `crates/db-merge/src/push_docs_transport.rs:109`, `crates/db-merge/src/push_docs_transport.rs:157` |
-| Backfill expands each head into an ordered transitive DAG block list. | `crates/db-merge/src/push_docs_transport.rs:167`, `crates/db-merge/src/push_docs_common.rs:8` |
-| Existing-doc replay sends blocks sequentially per document and stops that document on peer rejection or connection-like send failure. | `crates/db-merge/src/push_docs_transport.rs:199`, `crates/db-merge/src/push_docs_transport.rs:222` |
-| Retry recomputes a single document's latest heads and DAG blocks from storage rather than trusting a stale in-memory queue. | `crates/db-merge/src/push_docs_transport.rs:374`, `crates/db-merge/src/push_docs_transport.rs:416` |
+| Existing-doc replay waits for the target connection before it begins. | `crates/db/src/merge/push_docs_transport.rs:49` |
+| Backfill enumerates local document ids from the datastore, then loads the latest composite heads. | `crates/db/src/merge/push_docs_transport.rs:109`, `crates/db/src/merge/push_docs_transport.rs:157` |
+| Backfill expands each head into an ordered transitive DAG block list. | `crates/db/src/merge/push_docs_transport.rs:167`, `crates/db/src/merge/push_docs_common.rs:8` |
+| Existing-doc replay sends blocks sequentially per document and stops that document on peer rejection or connection-like send failure. | `crates/db/src/merge/push_docs_transport.rs:199`, `crates/db/src/merge/push_docs_transport.rs:222` |
+| Retry recomputes a single document's latest heads and DAG blocks from storage rather than trusting a stale in-memory queue. | `crates/db/src/merge/push_docs_transport.rs:374`, `crates/db/src/merge/push_docs_transport.rs:416` |
 | Live writes push the full document DAG to every matching replicator. | `crates/p2p/src/sync/coordinator/broadcast.rs:149`, `crates/p2p/src/sync/coordinator/broadcast.rs:179` |
 | Ordered live pushes stop on timeout or connection-like transport error and report a per-doc failure for retry. | `crates/p2p/src/sync/coordinator/broadcast.rs:436`, `crates/p2p/src/sync/coordinator/broadcast.rs:453`, `crates/p2p/src/sync/coordinator/broadcast.rs:238` |
